+++ include/promclient/promclient.h
@@ -1,0 +1,10 @@
+// Copyright 2017 Stefano Pogliani <stefano@spogliani.net>
+#ifndef PROMCLIENT_PROMCLIENT_H_
+#define PROMCLIENT_PROMCLIENT_H_
+
+// This is just an include wrapper to make all needed headers
+// available to library users.
+#include "promclient/counter.h"
+#include "promclient/internal/builder_counter.h"
+
+#endif  // PROMCLIENT_PROMCLIENT_H_